requirements.txt
flask
gunicorn
requests
openai
python-dotenv